move sign in automation to separate file
